further setup

+++ remove.php
@@ -1,13 +1,21 @@
 <?php
-$search = $_POST['username'];
-include "db.php";
+$password = $_POST['password'];
+$hash = ;
 
-$sql = "DELETE FROM twitchStreamers WHERE username like '%$search%'";
+if (password_verify($password, $hash)){
 
-if ($conn->query($sql) == TRUE) {
-    echo "Record deleted successfully";
+    include "db.php";
+    $search = $_POST['username'];
+
+    $sql = "DELETE FROM twitchStreamers WHERE username like '%$search%'";
+
+    if ($conn->query($sql) == TRUE) {
+        echo "Record deleted successfully";
+    } else {
+        echo "Error Deleting record: " .$conn->error;
+    }
+    $conn->close();
 } else {
-    echo "Error Deleting record: " .$conn->error;
+    echo 'Invalid Password!';
 }
-$conn->close();
 ?>
